requirements.txt
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-PyMongo==3.0.1
tensorflow==2.19.0
keras==3.9.1
numpy==2.1.3
pandas==2.2.3
matplotlib==3.10.1
scikit-learn==1.6.1
seaborn==0.13.2
joblib==1.4.2
xgboost==3.0.0
lime==0.2.0.1
shap==0.47.1
pillow==11.1.0
scikit-image==0.25.2
imbalanced-learn==0.13.0
PyYAML==6.0.2
Werkzeug==3.1.3
gunicorn==21.2.0
protobuf==5.29.4
python-dateutil==2.9.0.post0
six==1.17.0
requests==2.32.3
MarkupSafe==3.0.2
itsdangerous==2.2.0
Jinja2==3.1.6
bcrypt==4.3.0
dnspython==2.7.0
typing_extensions==4.13.0
uuid==1.30
